[Bug] Support modify non physical column comment
